refactor: rename origin-management module
